Add AI behavior trees and scripts
Adds behavior-tree assets and a collection of AI scripts to implement NPC behaviors and mechanics. New files include AiTreeLogic assets (AiLogic1, Claude, TreeLogicRat subgraphs), multiple behavior-tree node/action/condition scripts under "BT Creater By CS" and root-level AI condition/action/state classes (hunger, fleeing, perception, scared-of, etc.), enums, and a state machine. Also introduces Mechanics (AfterAttackAction, BeforeAttackAction, FeedOnKill, FleeBehavior, HungerManager, NotHungerCondition) and vision/AI updates. Modified files: Mechanics/Food.cs, Mechanics/HungerSystem.cs, Mechanics/RandomMovement.cs, RaicastSystem/AdaptiveVisionSystem.cs, RaicastSystem/EnemyAI.cs, and RandomoveAction.cs. These changes implement perception, hunger, random movement and flee behaviors and wire them into the behavior-tree assets for NPC AI.

+++ Mechanics/Food.cs
@@ -4,7 +4,7 @@ using UnityEngine.Events;
 public class Food : MonoBehaviour
 {
     [Header("Food Settings")]
-    [SerializeField] float satietyValue = 25f; // сколько сытости даёт эта пища
+    [SerializeField] int satietyValue = 25; // сколько сытости даёт эта пища
 
     [Header("Events")]
     public UnityEvent<float> OnFoodConsumed;
@@ -34,7 +34,7 @@ public class Food : MonoBehaviour
     /// </summary>
     public void SetSatietyValue(float value)
     {
-        satietyValue = Mathf.Max(0, value);
+        satietyValue = Mathf.Max(0, (int)value);
     }
 
     /// <summary>
